requirements.txt
openai
python-dotenv
customtkinter
httpx
certifi
